+++ Model/Wiki.php
@@ -9,6 +9,7 @@ use Kanboard\Core\Controller\AccessForbiddenException;
 use Kanboard\Model\UserModel;
 use SimpleValidator\Validator;
 use SimpleValidator\Validators;
+use Kanboard\Plugin\Wiki\Job\WikiEventJob;
 
 /**
  * Wiki
