+++ Test/Model/Wiki.php
@@ -40,7 +40,10 @@ class WikiPageTest extends Base
 
         $wikimodel = new Wiki($this->container);
         // create wiki pages
-        $this->assertEquals(1, $wikimodel->createpage($project['id'], "Security", "Some content", '2015-01-01'), 'Failed to a create wiki page on project');
+
+        $titleValue = "Security";
+        $contentValue = "Some content";
+        $this->assertEquals(1, $wikimodel->createpage($project['id'], $titleValue, $contentValue, '2015-01-01'), 'Failed to a create wiki page on project');
         $this->assertEquals(2, $wikimodel->createpage($project['id'], "Conventions", 'More content'), 'Failed to an additional create wiki page on project');
 
         // grab editions for first wiki page
@@ -48,8 +51,8 @@ class WikiPageTest extends Base
         $this->assertEmpty($editions);
 
         $values = [
-            'title' => "Security",
-            'content' => "Some content",
+            'title' => $titleValue,
+            'content' => $contentValue,
         ];
 
         // create wiki page edition
